Update course stageII

+++ server/controllers/Section.js
@@ -76,14 +76,54 @@ exports.updateSection = async (req, res) => {
 // DELETE a section
 exports.deleteSection = async (req, res) => {
 	try {
-		//HW -> req.params -> test
-		const { sectionId } = req.body;
+
+		const { sectionId, courseId }  = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+
+		//delete sub section
+		await SubSectionModal.deleteMany({_id: {$in: section.subSection}});
+
 		await Section.findByIdAndDelete(sectionId);
-		//HW -> Course ko bhi update karo
+
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+
 		res.status(200).json({
-			success: true,
-			message: "Section deleted",
+			success:true,
+			message:"Section deleted",
+			data:course
 		});
+
+
+
+
+
+	// 	//HW -> req.params -> test
+	// 	const { sectionId } = req.params;
+	// 	await Section.findByIdAndDelete(sectionId);
+	// 	//HW -> Course ko bhi update karo
+	// 	res.status(200).json({
+	// 		success: true,
+	// 		message: "Section deleted",
+	// 	});
 	} catch (error) {
 		console.error("Error deleting section:", error);
 		res.status(500).json({
